Initial Commit
Setup machinery workers and the CLI that starts said workers. Also added
the GORM DB ORM library and a global DB variable that can be imported
from all packages. Made the Broker type which acts as a wrapper for
sending async tasks using machinery. Added config interfaces and types
that unmarshal the config.json that should be placed in the root
directory and allow access to this unmarshalled struct